bulk_deliver/requirements.txt
fastapi
uvicorn
pydantic-settings
python-multipart
openpyxl
azure-data-tables
structlog
python-dotenv
